+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/mailer_transports.php
@@ -19,7 +19,6 @@ use Symfony\Component\Mailer\Bridge\MailerSend\Transport\MailerSendTransportFact
 use Symfony\Component\Mailer\Bridge\Mailgun\Transport\MailgunTransportFactory;
 use Symfony\Component\Mailer\Bridge\Mailjet\Transport\MailjetTransportFactory;
 use Symfony\Component\Mailer\Bridge\MailPace\Transport\MailPaceTransportFactory;
-use Symfony\Component\Mailer\Bridge\OhMySmtp\Transport\OhMySmtpTransportFactory;
 use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkTransportFactory;
 use Symfony\Component\Mailer\Bridge\Sendgrid\Transport\SendgridTransportFactory;
 use Symfony\Component\Mailer\Bridge\Sendinblue\Transport\SendinblueTransportFactory;
@@ -89,10 +88,6 @@ return static function (ContainerConfigurator $container) {
             ->tag('mailer.transport_factory')
 
         ->set('mailer.transport_factory.sendinblue', SendinblueTransportFactory::class)
-            ->parent('mailer.transport_factory.abstract')
-            ->tag('mailer.transport_factory')
-
-        ->set('mailer.transport_factory.ohmysmtp', OhMySmtpTransportFactory::class)
             ->parent('mailer.transport_factory.abstract')
             ->tag('mailer.transport_factory')