+++ src/Symfony/Component/Mailer/Tests/Exception/UnsupportedSchemeExceptionTest.php
@@ -20,7 +20,6 @@ use Symfony\Component\Mailer\Bridge\Mailchimp\Transport\MandrillTransportFactory
 use Symfony\Component\Mailer\Bridge\MailerSend\Transport\MailerSendTransportFactory;
 use Symfony\Component\Mailer\Bridge\Mailgun\Transport\MailgunTransportFactory;
 use Symfony\Component\Mailer\Bridge\Mailjet\Transport\MailjetTransportFactory;
-use Symfony\Component\Mailer\Bridge\OhMySmtp\Transport\OhMySmtpTransportFactory;
 use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkTransportFactory;
 use Symfony\Component\Mailer\Bridge\Sendgrid\Transport\SendgridTransportFactory;
 use Symfony\Component\Mailer\Bridge\Sendinblue\Transport\SendinblueTransportFactory;
@@ -42,7 +41,6 @@ final class UnsupportedSchemeExceptionTest extends TestCase
             MailgunTransportFactory::class => false,
             MailjetTransportFactory::class => false,
             MandrillTransportFactory::class => false,
-            OhMySmtpTransportFactory::class => false,
             PostmarkTransportFactory::class => false,
             SendgridTransportFactory::class => false,
             SendinblueTransportFactory::class => false,
@@ -71,7 +69,6 @@ final class UnsupportedSchemeExceptionTest extends TestCase
         yield ['mailgun', 'symfony/mailgun-mailer'];
         yield ['mailjet', 'symfony/mailjet-mailer'];
         yield ['mandrill', 'symfony/mailchimp-mailer'];
-        yield ['ohmysmtp', 'symfony/oh-my-smtp-mailer'];
         yield ['postmark', 'symfony/postmark-mailer'];
         yield ['sendgrid', 'symfony/sendgrid-mailer'];
         yield ['sendinblue', 'symfony/sendinblue-mailer'];