+++ src/Symfony/Component/Workflow/Registry.php
@@ -18,7 +18,7 @@ use Symfony\Component\Workflow\SupportStrategy\WorkflowSupportStrategyInterface;
  * @author Fabien Potencier <fabien@symfony.com>
  * @author Grégoire Pineau <lyrixx@lyrixx.info>
  *
- * @internal since Symfony 6.2. Inject the workflow where you need it.
+ * @internal
  */
 class Registry
 {