+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+7.0
+---
+
+ * Remove the `Twig_Environment` autowiring alias, use `Twig\Environment` instead
+ * Remove option `twig.autoescape`; create a class that implements your escaping strategy
+   (check `FileExtensionEscapingStrategy::guess()` for inspiration) and reference it using
+   the `twig.autoescape_service` option instead
+
 6.3
 ---