+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
  * Remove `AbstractSurrogate::$phpEscapeMap`
  * Remove `HttpKernelInterface::MASTER_REQUEST`
  * Remove `terminate_on_cache_hit` option from `HttpCache`
+ * Require explicit argument when calling `ConfigDataCollector::setKernel()`, `RouterListener::setCurrentRequest()`
 
 6.4
 ---